Modify kernel load address to 0x80200000
When the default bios is used it is loaded at 0x80000000.
Currently OpenSBI is the default bios and reserves a couple
hundred kilobytes for itself. A 2MiB alignment is used for
the new kernel load address simply because it's common in
the Linux world for RISC-V. In reality the kernel load address
could be immediately after OpenSBI. Since the size of OpenSBI
changes using 2MiB should give us plenty of room.